pushed time_check and np_check out to utils.py
